Move frontmatter analysis code into its own module
